For 'fastboot oem set-storage', support several storage type now.
And this command can be used in user build now.
And does not support storage type "virtual_media" now.

User need to use this command to change the default install device,
such as load kernelflinger from USB disk and enter fastboot mode.
The old command will ignore the currently used storage, and return
failed if can't find new storage.
The new code will check whether the currently used storage meet the
required storage type, and return successful if meet.

After this patch, then:
1. The device has eMMC, and use RPMB, and the device state stored in
RPMB is set to locked.
2. Boot the device from USB disk, then will use simulate RPMB in teedata
partition of this USB disk, and the device state stored in simulate
RPMB is set to unlocked.
3. After use 'fastboot oem set-storage emmc', then will init RPMB again,
and refresh the device state. So the device state will change to
locked.

If you flash a bootloader to a device with wrong RPMB key, then the
device will be in locked state, and you can't boot the same bootloader
from a USB disk to unlock the device. You should use a bootloader with
correct RPMB key, or build a bootloader which does not use physical RPMB
to try to unlock the device.

Also publish a fastboot var 'boot-device' to show current boot device.
In sometimes, the boot-device var is too long and will be cut.

At last, fix some error and warning of checkpatch.pl.

Change-Id: I29d5da812d940a193adcc12b598782fc5088b5b7
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-66241
Signed-off-by: Ming Tan <ming.tan@intel.com>
Reviewed-on: https://android.intel.com:443/631430