android: handover_kernel() - fix setup_header copy
According to the Linux Documentation/x86/boot.rst:

The memory for struct boot_params should be allocated and initialized
to all zero. Then the setup header from offset 0x01f1 of kernel image
on should be loaded into struct boot_params and examined. The end of
setup header can be calculated as follow:
    0x0202 + byte value at offset 0x0201

Signed-off-by: JianFeng,Zhou <jianfeng.zhou@intel.com>